% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{buildCacheKey}
\alias{buildCacheKey}
\title{Construct a unique cache key for a request}
\usage{
buildCacheKey(url, query = NULL, body = NULL, extras = c())
}
\arguments{
\item{url}{character request URL}

\item{query}{Optional query parameters for the request}

\item{body}{Optional request body}

\item{extras}{character Optional additional annotations to include in the
cache key.}
}
\value{
Character value, starting with \code{url} and including hashed query
and body values if provided, to be used as the cache key for this request.
}
\description{
This function encapsulates the logic of making a cache key, allowing other
code or libraries to access the HTTP cache programatically.
}

