% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache-api}
\alias{cache-api}
\alias{getCache}
\alias{hitCache}
\alias{setCache}
\title{HTTP Cache API}
\usage{
hitCache(key)

getCache(key)

setCache(key, value)
}
\arguments{
\item{key}{character, typically a URL or similar}

\item{value}{For \code{setCache}, an R object to set in the cache for
\code{key}.}
}
\value{
\code{hitCache} returns logical whether \code{key} exists in the
cache. \code{getCache} returns the value stored in the cache, or \code{NULL}
if there is nothing cached. \code{setCache} is called for its side effects.
}
\description{
These functions provide access to what's stored in the cache.
}

