% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache-management}
\alias{cache-management}
\alias{cacheOff}
\alias{cacheOn}
\alias{clearCache}
\title{Manage the HTTP cache}
\usage{
cacheOn()

cacheOff()

clearCache()
}
\value{
Nothing. Functions are run for their side effects.
}
\description{
These functions turn the cache on and off and clear the contents of the
query cache.
}

