% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{cacheLogSummary}
\alias{cacheLogSummary}
\title{Summarize cache performance from a log}
\usage{
cacheLogSummary(logdf)
}
\arguments{
\item{logdf}{A logging data.frame, as loaded by \code{link{loadLogfile}}.}
}
\value{
A list containing counts of cache hit/set/drop events, plus a
cache hit rate.
}
\description{
Summarize cache performance from a log
}

