% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-manager.R
\name{ContextManager}
\alias{ContextManager}
\alias{contextManager}
\title{Context managers}
\usage{
ContextManager(enter = function() { }, exit = function() { },
  error = NULL, as = NULL)
}
\arguments{
\item{enter}{function to run before doing things}

\item{exit}{function to run after doing things}

\item{error}{optional function to run if an error is thrown}

\item{as}{character optional way to specify a default name for assinging
the return of the enter function.}
}
\value{
an S3 class "contextManager" object
}
\description{
Context managers
}
\seealso{
\link{with-context-manager}
}

