% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dropCache}
\alias{dropCache}
\alias{dropOnly}
\alias{dropPattern}
\title{Invalidate cache}
\usage{
dropCache(x)

dropOnly(x)

dropPattern(x)
}
\arguments{
\item{x}{character URL or regular expression}
}
\value{
Nothing. Functions are run for their side effects.
}
\description{
These functions let you control cache invalidation. \code{dropOnly}
invalidates cache only for the specified URL. \code{dropPattern} uses
regular expression matching to invalidate cache. \code{dropCache} is a
convenience wrapper around \code{dropPattern} that invalidates cache for
any resources that start with the given URL.
}
