% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{halt}
\alias{halt}
\title{Stop, log, and no call}
\usage{
halt(..., call. = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{stop}}

\item{call.}{logical: print the call? Default is \code{FALSE}, unlike \code{stop}}
}
\value{
Nothing. Raises an error.
}
\description{
Wrapper around \code{\link[base:stop]{base::stop()}} that logs the error message and then stops
with call.=FALSE by default.
}
