% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{startLog}
\alias{startLog}
\title{Enable logging}
\usage{
startLog(filename = "", append = FALSE)
}
\arguments{
\item{filename}{character: a filename/path where the log can be written out.
If \code{""}, messages will print to stdout (the screen). See \code{\link[base:cat]{base::cat()}}.}

\item{append}{logical: if the file already exists, append to it? Default
is \code{FALSE}, and if not in append mode, if the \code{filename} exists,
it will be deleted.}
}
\value{
Nothing.
}
\description{
Enable logging
}
