% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-manager.R
\name{temp.options}
\alias{temp.option}
\alias{temp.options}
\title{Set some global options temporarily}
\usage{
temp.options(...)

temp.option(...)
}
\arguments{
\item{...}{named options to set}
}
\value{
an S3 class "contextManager" object
}
\description{
Set some global options temporarily
}
\seealso{
\link{with-context-manager} \link{ContextManager}
}

