% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-manager.R
\name{with-context-manager}
\alias{with-context-manager}
\alias{with.contextManager}
\title{Context manager's "with" method}
\usage{
\method{with}{contextManager}(data, expr, ...)
}
\arguments{
\item{data}{\code{\link{contextManager}}}

\item{expr}{code to evaluate within that context}

\item{...}{additional arguments. One additional supported argument is "as",
which lets you assign the return of your "enter" function to an object you
can access.}
}
\value{
Nothing.
}
\description{
Context manager's "with" method
}
\seealso{
\link{ContextManager}
}

