% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-cache.R
\name{saveCache}
\alias{saveCache}
\alias{loadCache}
\title{Save and load cache state}
\usage{
saveCache(file)

loadCache(file)
}
\arguments{
\item{file}{character file path to write the cache data to, in \code{.rds} format}
}
\value{
Nothing; called for side effects.
}
\description{
Warm your query cache from a previous session by saving out the cache and
loading it back in.
}
