% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{uncached}
\alias{uncached}
\title{Context manager to temporarily turn cache off if it is on}
\usage{
uncached(...)
}
\arguments{
\item{...}{Things to evaluate with caching off}
}
\value{
Whatever ... returns.
}
\description{
If you don't want to store the response of a GET request in the cache,
wrap it in \code{uncached()}. It will neither read from nor write to cache.
However, \code{uncached} will not invalidate cache records, if present.
}
\examples{
uncached(GET("http://httpbin.org/get"))
}

